fix(vitest-angular): setup separate import for configuring snapshots
