+++ packages/vitest-angular/tsconfig.lib.json
@@ -5,6 +5,6 @@
     "declaration": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts", "setup-zone.ts"],
+  "include": ["src/**/*.ts", "setup-zone.ts", "setup-snapshots.ts"],
   "exclude": ["src/**/*.spec.ts", "src/**/*.test.ts"]
 }
